backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
pydantic-settings==2.8.1
pymongo==4.11.2
bcrypt==4.0.1